requirements.txt
scrapy==1.4.0
flask==0.12.2
pymongo==2.8.1
MongoAlchemy==0.19
flask_mongoalchemy==0.7.2
beautifulsoup4==4.6.0
python-dateutil
dateparser==0.7.1
requests==2.22.0
pypiwin32